Add tasks summary concern
