+++ app/helpers/dorsale/flyboy/application_helper.rb
@@ -25,15 +25,6 @@ module Dorsale
         render "dorsale/flyboy/tasks/tasks_for_taskable", tasks: tasks, taskable: taskable
       end
 
-      def setup_tasks_summary
-        tasks = current_user_scope.tasks.where("(owner_id is null and owner_type is null) or (owner_id = ? and owner_type = ?)", current_user.id, current_user.class.name)
-        @delayed_tasks   = tasks.delayed
-        @today_tasks     = tasks.today
-        @tomorrow_tasks  = tasks.tomorrow
-        @this_week_tasks = tasks.this_week
-        @next_week_tasks = tasks.next_week
-      end
-
       def show_tasks_summary
         render "dorsale/flyboy/tasks/summary"
       end
